librerias
Seguimos mejorando y comentamos código innecesario

+++ js/imc.js
@@ -265,20 +265,36 @@ function agregarPaciente() {
 
 function borrarUltimoPaciente() {
   
-  arrayPersonasRegistradas = JSON.parse(localStorage.getItem("personasRegistradas")) || [];
 
-  arrayPersonasRegistradas.pop();
-  const listaTr = document.querySelectorAll("tr");
-  listaTr.forEach((elemento, i) => {
-    if (i == listaTr.length - 1 && i != 0) {
-      elemento.remove();
+
+  Swal.fire({
+    title: '¿Desea eliminar el último paciente ingresado?',
+    showDenyButton: true,
+    showCancelButton: true,
+    confirmButtonText: 'Aceptar',
+    denyButtonText: `Cancelar`,
+  }).then((result) => {
+    if (result.isConfirmed) {
+      Swal.fire('Eliminado!', '', 'success')
+      arrayPersonasRegistradas = JSON.parse(localStorage.getItem("personasRegistradas")) || [];
+
+      arrayPersonasRegistradas.pop();
+      const listaTr = document.querySelectorAll("tr");
+      listaTr.forEach((elemento, i) => {
+        if (i == listaTr.length - 1 && i != 0) {
+          elemento.remove();
+        }
+      });
+
+      localStorage.setItem(
+        "personasRegistradas",
+        JSON.stringify(arrayPersonasRegistradas)
+      );
+    } else if (result.isDenied) {
+      Swal.fire('Paciente NO eliminado', '', '')
     }
-  });
-
-  localStorage.setItem(
-    "personasRegistradas",
-    JSON.stringify(arrayPersonasRegistradas)
-  );
+  })
+  
 }
 
 function modificarTitulo() {
@@ -287,17 +303,17 @@ function modificarTitulo() {
   titulo.style.textAlign = "center";
 }
 
-const definirComposicionCorporal = (imc) => {
-  if (imc < 18.5) {
-    alert("Composición corporal: Peso inferior al normal");
-  } else if (imc > 18.5 && imc < 24.9) {
-    alert("Composición corporal: Peso normal");
-  } else if (imc > 25 && imc < 29.9) {
-    alert("Composición corporal: Peso superior al normal");
-  } else {
-    alert("Composición corporal: Obesidad");
-  }
-};
+// const definirComposicionCorporal = (imc) => {
+//   if (imc < 18.5) {
+//     alert("Composición corporal: Peso inferior al normal");
+//   } else if (imc > 18.5 && imc < 24.9) {
+//     alert("Composición corporal: Peso normal");
+//   } else if (imc > 25 && imc < 29.9) {
+//     alert("Composición corporal: Peso superior al normal");
+//   } else {
+//     alert("Composición corporal: Obesidad");
+//   }
+// };
 
 function verificarDato(dato) {
   while (dato == 0 || Number.isNaN(dato)) {
@@ -314,30 +330,30 @@ function calcularIMC(peso, altura) {
   return Number((peso / Math.pow(altura, 2)).toFixed(2));
 }
 
-function mostrar() {
-  dni = verificarDato(Number(prompt("Ingrese su DNI sin guiones ni puntos")));
-  nombre = datoSinEspacio(capitalizarPrimeraLetra(prompt("Ingrese su nombre")));
-  apellido = datoSinEspacio(
-    capitalizarPrimeraLetra(prompt("Ingrese su apellido"))
-  );
-  estatura = verificarDato(Number(prompt("Ingrese su estatura en metros")));
-  peso = verificarDato(Number(prompt("Ingrese su peso en KG")));
-  imc = calcularIMC(peso, estatura);
-  alert("Su IMC es: " + imc);
-  definirComposicionCorporal(imc);
-  const objeto1 = new Persona(dni, nombre, apellido, estatura, peso, imc);
-  personasRegistradasDia.push(objeto1);
-  if (confirm("Quiere verificar otra persona")) {
-    mostrar();
-  } else {
-    console.log(
-      `Estás son las personas registradas hoy en día ${JSON.stringify(
-        personasRegistradasDia
-      )}`
-    );
-    console.table(personasRegistradasDia);
-  }
-}
+// function mostrar() {
+//   dni = verificarDato(Number(prompt("Ingrese su DNI sin guiones ni puntos")));
+//   nombre = datoSinEspacio(capitalizarPrimeraLetra(prompt("Ingrese su nombre")));
+//   apellido = datoSinEspacio(
+//     capitalizarPrimeraLetra(prompt("Ingrese su apellido"))
+//   );
+//   estatura = verificarDato(Number(prompt("Ingrese su estatura en metros")));
+//   peso = verificarDato(Number(prompt("Ingrese su peso en KG")));
+//   imc = calcularIMC(peso, estatura);
+//   alert("Su IMC es: " + imc);
+//   definirComposicionCorporal(imc);
+//   const objeto1 = new Persona(dni, nombre, apellido, estatura, peso, imc);
+//   personasRegistradasDia.push(objeto1);
+//   if (confirm("Quiere verificar otra persona")) {
+//     mostrar();
+//   } else {
+//     console.log(
+//       `Estás son las personas registradas hoy en día ${JSON.stringify(
+//         personasRegistradasDia
+//       )}`
+//     );
+//     console.table(personasRegistradasDia);
+//   }
+// }
 
 function buscarImcMenores() {
   
@@ -471,93 +487,93 @@ function buscarImcMayores() {
 
 }
 
-function buscarImcMaximo() {
+// function buscarImcMaximo() {
 
-  arrayPersonasRegistradas = JSON.parse(localStorage.getItem("personasRegistradas")) || [];
+//   arrayPersonasRegistradas = JSON.parse(localStorage.getItem("personasRegistradas")) || [];
 
-  if (arrayPersonasRegistradas) {
-    const arrayImc = arrayPersonasRegistradas.map((persona) => persona.imc);
-    console.log(arrayImc)
-    imcMaximo = Math.max(...arrayImc);
-    console.log(imcMaximo)
-    indices = []
-    idMax = arrayImc.indexOf(imcMaximo) + 1;
-    while (imcMaximo != -1) {
-      indices.push(imcMaximo);
-      idMax = array.indexOf(imcMaximo, idMax + 1);
-    }
-    console.log(indices)
-    console.log
-    personaImcMaximo = arrayPersonasRegistradas.find(
-      (persona) => persona.imc == imcMaximo
-    );
+//   if (arrayPersonasRegistradas) {
+//     const arrayImc = arrayPersonasRegistradas.map((persona) => persona.imc);
+//     console.log(arrayImc)
+//     imcMaximo = Math.max(...arrayImc);
+//     console.log(imcMaximo)
+//     indices = []
+//     idMax = arrayImc.indexOf(imcMaximo) + 1;
+//     while (imcMaximo != -1) {
+//       indices.push(imcMaximo);
+//       idMax = array.indexOf(imcMaximo, idMax + 1);
+//     }
+//     console.log(indices)
+//     console.log
+//     personaImcMaximo = arrayPersonasRegistradas.find(
+//       (persona) => persona.imc == imcMaximo
+//     );
 
-    eliminarFilas();
+//     eliminarFilas();
 
-    tbody = document.querySelector("tbody");
+//     tbody = document.querySelector("tbody");
 
-    if (personaImcMaximo.imc < 25) {
-      color = "green";
-    } else if (personaImcMaximo.imc > 25 && personaImcMaximo.imc < 30) {
-      color = "yellow";
-    } else {
-      color = "red";
-    }
+//     if (personaImcMaximo.imc < 25) {
+//       color = "green";
+//     } else if (personaImcMaximo.imc > 25 && personaImcMaximo.imc < 30) {
+//       color = "yellow";
+//     } else {
+//       color = "red";
+//     }
 
-    tr = document.createElement("tr");
+//     tr = document.createElement("tr");
 
-    tr.innerHTML = ` 
-    <th scope="col">${idMax}</th>
-    <td>${personaImcMaximo.dni}</td>
-    <td>${personaImcMaximo.nombre}</td>
-    <td>${personaImcMaximo.apellido}</td>
-    <td>${personaImcMaximo.estatura}</td>
-    <td>${personaImcMaximo.peso}</td>
-    <td style="color:${color}">${personaImcMaximo.imc}</td>
-    `;
-      tbody.append(tr);
-  }
-}
+//     tr.innerHTML = ` 
+//     <th scope="col">${idMax}</th>
+//     <td>${personaImcMaximo.dni}</td>
+//     <td>${personaImcMaximo.nombre}</td>
+//     <td>${personaImcMaximo.apellido}</td>
+//     <td>${personaImcMaximo.estatura}</td>
+//     <td>${personaImcMaximo.peso}</td>
+//     <td style="color:${color}">${personaImcMaximo.imc}</td>
+//     `;
+//       tbody.append(tr);
+//   }
+// }
 
-function buscarImcMinimo() {
-  arrayPersonasRegistradas = JSON.parse(
-    localStorage.getItem("personasRegistradas")
-  );
-  if (arrayPersonasRegistradas) {
-    const arrayImc = arrayPersonasRegistradas.map((persona) => persona.imc);
-    imcMinimo = Math.min(...arrayImc);
-    id = arrayImc.indexOf(imcMinimo) + 1;
+//  function buscarImcMinimo() {
+//    arrayPersonasRegistradas = JSON.parse(
+//     localStorage.getItem("personasRegistradas")
+//   );
+//   if (arrayPersonasRegistradas) {
+//     const arrayImc = arrayPersonasRegistradas.map((persona) => persona.imc);
+//     imcMinimo = Math.min(...arrayImc);
+//     id = arrayImc.indexOf(imcMinimo) + 1;
     
-    personaImcMinimo = arrayPersonasRegistradas.find(
-      (persona) => persona.imc == imcMinimo
-    );
+//     personaImcMinimo = arrayPersonasRegistradas.find(
+//       (persona) => persona.imc == imcMinimo
+//     );
 
-    eliminarFilas();
+//     eliminarFilas();
     
-    tbody = document.querySelector("tbody");
+//     tbody = document.querySelector("tbody");
 
-    if (personaImcMinimo.imc < 25) {
-      color = "green";
-    } else if (personaImcMinimo.imc > 25 && personaImcMinimo.imc < 30) {
-      color = "yellow";
-    } else {
-      color = "red";
-    }
+//     if (personaImcMinimo.imc < 25) {
+//       color = "green";
+//     } else if (personaImcMinimo.imc > 25 && personaImcMinimo.imc < 30) {
+//       color = "yellow";
+//     } else {
+//       color = "red";
+//     }
 
-    tr = document.createElement("tr");
+//     tr = document.createElement("tr");
 
-    tr.innerHTML = ` 
-    <th scope="col">${id}</th>
-    <td>${personaImcMinimo.dni}</td>
-    <td>${personaImcMinimo.nombre}</td>
-    <td>${personaImcMinimo.apellido}</td>
-    <td>${personaImcMinimo.estatura}</td>
-    <td>${personaImcMinimo.peso}</td>
-    <td style="color:${color}">${personaImcMinimo.imc}</td>
-    `;
-      tbody.append(tr);
-  }
-}
+//     tr.innerHTML = ` 
+//     <th scope="col">${id}</th>
+//     <td>${personaImcMinimo.dni}</td>
+//     <td>${personaImcMinimo.nombre}</td>
+//     <td>${personaImcMinimo.apellido}</td>
+//     <td>${personaImcMinimo.estatura}</td>
+//     <td>${personaImcMinimo.peso}</td>
+//     <td style="color:${color}">${personaImcMinimo.imc}</td>
+//     `;
+//       tbody.append(tr);
+//    }
+//  }
 
 implementarDom();
